+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 1.1.14
+- Add cookies to the cookie manager
+- Resize icon to 40px * 40px
+- Fix line item attributes
+
+# 1.1.13
+- Include vendor folder in Shopware store releases
+
+# 1.1.12
+- Update doc path
+
+# 1.1.11
+- Add documentation
+
 # 1.1.10
 - Stop responding with server errors when orders are not found